Separate gamemodes in a specific module
