+++ src/game/mod.rs
@@ -1,0 +1,2 @@
+pub mod as_human;
+pub mod as_ia;
